# Server: enable lookup decoding