--- a/examples/server/tests/features/steps/steps.py
+++ b/examples/server/tests/features/steps/steps.py
@@ -3,6 +3,7 @@ import collections
 import json
 import os
 import re
+import signal
 import socket
 import subprocess
 import sys
@@ -873,6 +874,8 @@ async def request_completion(prompt,
                                 headers=headers,
                                 timeout=3600) as response:
             if expect_api_error is None or not expect_api_error:
+                if 'DEBUG' in os.environ and os.environ['DEBUG'] == 'ON' and response.status != 200:
+                    print(f"Unexpected bad HTTP response: {response.status}")
                 assert response.status == 200
                 assert response.headers['Access-Control-Allow-Origin'] == origin
                 return await response.json()
@@ -1260,8 +1263,7 @@ def start_server_background(context):
         server_args.extend(['--ubatch-size', context.n_ubatch])
     if context.n_gpu_layer:
         server_args.extend(['--n-gpu-layers', context.n_gpu_layer])
-    if context.draft is not None:
-        server_args.extend(['--draft', context.draft])
+    server_args.extend(['--draft', context.draft if context.draft is not None else 0])
     if context.server_continuous_batching:
         server_args.append('--cont-batching')
     if context.server_embeddings:
@@ -1303,6 +1305,14 @@ def start_server_background(context):
         'stdout': subprocess.PIPE,
         'stderr': subprocess.PIPE
     }
+
+    if context.server_process is not None:
+        if os.name == 'nt':
+            interrupt = signal.CTRL_C_EVENT
+        else:
+            interrupt = signal.SIGINT
+        context.server_process.send_signal(interrupt)
+
     context.server_process = subprocess.Popen(
         [str(arg) for arg in [context.server_path, *server_args]],
         **pkwargs)