… examples/lookup/README.md
… examples/server/server.cpp