This PR aims to enable lookup decoding for the llama.cpp server in the same way as it is used in `examples/lookup`, see https://github.com/ggerganov/llama.cpp/pull/5479 . To recapitulate, the implementation tries to guess the next few tokens that will be generated using simple text statistics. I *think* the current implementation works but it is difficult to properly benchmark. The intended way for it to work is:

* Start with empty context cache, empty dynamic cache (unless the user provides one), and static cache loaded from file.
* When generating tokens, try to continue with context cache, validate with static cache.
* If that fails, try to continue with dynamic cache, validate with static cache.
* If that fails, try to continue with static cache.
* When new tokens are generated, use them to update the context cache.
* When a generation is finished, update the dynamic cache with the context cache, then empty the context cache.
* On server shutdown, save dynamic cache to file if the user provided a path.

These are the results I get from `examples/server/bench.py` using an RTX 4090 and various static lookup caches and an initially empty dynamic cache:

| Model                          | Static lookup cache | Iterations master |    Iterations PR | Speedup |
| ------------------------------ | ------------------- |  ---------------: | ---------------: | ------: |
| Phi-2 3b q4_0                  | None                |               274 |              365 |    1.33 |
| Phi-2 3b q4_0                  | Wikitext 103        |               274 |              361 |    1.32 |
| Phi-2 3b q4_0                  | Mistral 1.64e8      |               274 |              354 |    1.29 |
| LLaMA 2 7b q4_0                | None                |               148 |              256 |    1.73 |
| LLaMA 2 7b q4_0                | Wikitext 103        |               148 |              255 |    1.72 |
| LLaMA 2 7b q4_0                | Mistral 1.64e8      |               148 |              255 |    1.72 |

It does seem to provide a speedup but adding a static lookup cache does not seem to help (the caches are created either from Wikitext 103 or from 164 million tokens generated with Mistral q8_0). Assuming there are no bugs, what I think is happening is that the dataset for the benchmark (see server bench README) is very repetitive so using a static cache pulls the drafts away from these very repetitive patterns and reduces the speed. Also for Phi-2 in particular I think that I simply don't have enough input data for the static cache to get sufficiently precise text statistics (since it has a larger vocabulary size). Regarding the latter, I recently built a machine with 6x RTX 4090 so I think I will be able to significantly scale up the rate at which I can produce synthetic text (I was previously using 3x P40 and 1x RX 6800).

In this PR I also changed the interface of `llama_ngram_cache_load` to be more in line with the rest of llama.cpp; I'll maybe change the interface of some of the other functions as well.

Also: is it somehow possible to retrieve the tokens that were previously fed to the model? I'm currently manually tracking this in `server.cpp` but this adds the potential for error.