@phymbert 2024-04-22
Great work. As we discussed previously, servers' test coverage matters, and adding a new scenario in the test framework is mandatory.