@phymbert 2024-04-22
I was not aware, but this is not asserted in the parallel test suite AFAIK.

Also, I recall that each architecture generates different results.