@arnfaldur 2024-04-25
A research paper studying this exact technique was recently published and suggested for integration in an issue https://github.com/ggerganov/llama.cpp/issues/6813
I have been looking at the current implementation and trying to make it match the implementation in the paper.

- The short version of it is that there are no `static` or `dynamic` N-gram caches, only ones like you call `context`, generated on the fly.
- They use what they call a multi level N-gram. There are multiple caches. Each cache only considers N-grams of length N. When querying the multi level N-gram, the module with the longest N-grams that match the token suffix is the one used.
- They then generate K tokens using this multi level N-gram as a proposal to be validated by the LLM.
- Based on their ablation study, they suggest N = 5 and K = 7.

I haven't spent enough time reading `ngram-cache.cpp` and friends to tell how it works and how it differs, besides the persistance.