@JohannesGaessler 2024-04-25
I forgot: if you want to play around with the llama.cpp implementation, take a look at `lookup-stats`. It has the same interface as `perplexity` and can be used to estimate how many tokens tokens would be predicted on a pre-generated text (so you don't have to actually evaluate the model).