@JohannesGaessler 2024-04-25
>How much does this PR increase token generation? As far I am aware https://github.com/ggerganov/llama.cpp/pull/5479 had rather tiny speedup.

Should be something like 1.1-1.4 for natural language with an essentially empty context. For source code or summarization it's going to be a lot more. The numbers in the OP are indicative of the long-term speedup using similar prompts once the dynamic cache fills up.

>And when do you think this PR will be ready to be merged?

I'm aiming for the end of the week.