@JohannesGaessler 2024-04-28
Functionally I think everything is in order now. Unfortunately I think that it's currently not possible to get bit-for-bit identical results with lookup decoding since the results seem to change slightly when the batch size is varied, see https://github.com/ggerganov/llama.cpp/pull/6950 . For this reason there are no automated tests for lookup decoding that assert that the results do not change (because they do).