@JohannesGaessler 2024-04-28
I'm very sorry but it seems the numbers that I previously reported were incorrect. The speed I reported for "master" was actually the speed for this PR with `--draft 0`. However, this means that the numbers were still including the overhead associated with the lookup caches which is quite significant. These are the correct numbers for the most recent version:

| Model         | Static lookup cache | Slots | Iterations master | Iterations PR --draft 0 | Iterations PR --draft 5 | Speedup vs. master | Speedup vs. --draft 0 |
|---------------+---------------------+-------+-------------------+-------------------------+-------------------------+--------------------+-----------------------|
| Phi-2 3b q4_0 | None                |     1 |               549 |                     274 |                     363 |               0.66 |                  1.32 |
| Phi-2 3b q4_0 | None                |     2 |               947 |                     455 |                     599 |               0.63 |                  1.32 |
| Phi-2 3b q4_0 | None                |     4 |              1465 |                     704 |                     797 |               0.54 |                  1.13 |
| Phi-2 3b q4_0 | None                |     8 |              1856 |                     855 |                     900 |               0.48 |                  1.05 |

For Phi-2 on an RTX 4090 there is a regression relative to master because it is quite fast so the constant overhead per token is too large relative to the speedup. I'll investigate performance for larger models/slower hardware.