@JohannesGaessler 2024-04-28
Performance for LLaMA 3 70 on 3x RTX 4090 is looking much better:

| Model              | Static lookup cache | Slots | Iterations master | Iterations PR --draft 5 | Speedup vs. master |
|--------------------+---------------------+-------+-------------------+-------------------------+--------------------|
| LLaMA 3 70b q4_K_M | None                |     1 |                24 |                      44 |               1.83 |
| LLaMA 3 70b q4_K_M | WT 103              |     1 |                24 |                      42 |               1.75 |