@Green-Sky 2024-04-29
Regarding performance, it seems your hashes for the lookup table are of low quality. `std::hash<llama_token>` is the same as `std::hash<int32_t>`, which just returns the identity of the token. Also the standard containers are known to be not performing best in class, but that's a different issue. :)