@JohannesGaessler 2024-04-29
Prior to reading the hashing function blog post I wrote a simple implementation that just uses bit shifts and xors but that already results in much better performance:

| Model         | Static lookup cache | Slots | Iterations master | Iterations PR --draft 5 | Speedup vs. master |
|---------------+---------------------+-------+-------------------+-------------------------+--------------------|
| Phi-2 3b q4_0 | None                |     1 |               549 |                     634 |               1.15 |
| Phi-2 3b q4_0 | None                |     2 |               947 |                    1113 |               1.18 |
| Phi-2 3b q4_0 | None                |     4 |              1465 |                    1572 |               1.07 |
| Phi-2 3b q4_0 | None                |     8 |              1856 |                    1790 |               0.96 |
| Phi-2 3b q4_0 | WT 103              |     1 |               549 |                     643 |               1.17 |
| Phi-2 3b q4_0 | WT 103              |     2 |               947 |                    1098 |               1.16 |
| Phi-2 3b q4_0 | WT 103              |     4 |              1465 |                    1549 |               1.06 |
| Phi-2 3b q4_0 | WT 103              |     8 |              1856 |                    1766 |               0.95 |