@JohannesGaessler 2024-05-02
I've added a test for asserting that lookup decoding produces correct results. The sequences are the same for temperature 0 though the results are not going to be bit-for-bit identical. I've also investigated the performance for LLaMA 3 Instruct in more detail:

<details>
<summary>Results</summary>

| Model                       | GPU         | Static lookup cache | Slots | Seed | Previous runs | Iterations master | Iterations PR --draft 5 | Speedup vs. master |
|-----------------------------|-------------|---------------------|-------|------|---------------|-------------------|-------------------------|--------------------|
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | None                |     1 |   42 |             0 |               166 |                     182 |               1.10 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | None                |     2 |   42 |             0 |               278 |                     283 |               1.02 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | None                |     4 |   42 |             0 |               429 |                     367 |               0.86 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | None                |     8 |   42 |             0 |               531 |                     407 |               0.77 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   42 |             0 |               166 |                     183 |               1.10 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   42 |             0 |               278 |                     282 |               1.01 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   42 |             0 |               429 |                     355 |               0.83 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   -1 |             0 |               166 |                     182 |               1.10 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   -1 |             1 |               166 |                     186 |               1.12 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   -1 |             2 |               166 |                     204 |               1.23 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   -1 |             3 |               166 |                     208 |               1.25 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   -1 |             4 |               166 |                     215 |               1.30 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   -1 |             5 |               166 |                     219 |               1.32 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   -1 |             6 |               166 |                     222 |               1.34 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     1 |   -1 |             7 |               166 |                     222 |               1.34 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   -1 |             0 |               278 |                     285 |               1.03 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   -1 |             1 |               278 |                     283 |               1.02 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   -1 |             2 |               278 |                     309 |               1.11 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   -1 |             3 |               278 |                     315 |               1.13 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   -1 |             4 |               278 |                     324 |               1.17 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   -1 |             5 |               278 |                     326 |               1.17 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   -1 |             6 |               278 |                     329 |               1.18 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     2 |   -1 |             7 |               278 |                     333 |               1.20 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   -1 |             0 |               429 |                     363 |               0.85 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   -1 |             1 |               429 |                     353 |               0.82 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   -1 |             2 |               429 |                     370 |               0.86 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   -1 |             3 |               429 |                     378 |               0.88 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   -1 |             4 |               429 |                     378 |               0.88 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   -1 |             5 |               429 |                     383 |               0.89 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   -1 |             6 |               429 |                     383 |               0.89 |
| LLaMA 3 Instruct 8b q4_K_M  | 1x RTX 4090 | WT 103              |     4 |   -1 |             7 |               429 |                     388 |               0.90 |
| LLaMA 3 Instruct 70b q4_K_M | 2x RTX 4090 | None                |     1 |   42 |             0 |                28 |                      31 |               1.11 |
| LLaMA 3 Instruct 70b q4_K_M | 2x RTX 4090 | None                |     2 |   42 |             0 |                55 |                      57 |               1.04 |
| LLaMA 3 Instruct 70b q4_K_M | 2x RTX 4090 | None                |     4 |   42 |             0 |                96 |                      69 |               0.72 |
| LLaMA 3 Instruct 70b q4_K_M | 2x RTX 4090 | None                |     8 |   42 |             0 |               120 |                     OOM |                ??? |

</details>

The speedup between LLaMA 3 instruct 8b and 70b seems to be very similar. The current implementation is only faster for small numbers of slots since there is comparatively less benefit for adding more tokens to the batch if you're already at 8 tokens per batch without any speculative decoding. Successive, similar runs with different seeds but a carried over dynamic cache result in increasing performance over time, for a single slot the 8th run was ~1.2x faster than the first one.

From my side I would consider this PR ready to be merged if one last issue is resolved: whether n-gram lookup should be enabled or disabled by default. The default for the number of slots is 1 and for that case it is faster. However, due to the varying batch size it also causes nondeterministic results. I personally would tend more towards having n-gram lookup be disabled by default but do not have a strong opinion on it.