@JohannesGaessler 2024-05-09
I re-tested the performance on 1x RTX 4090 with CUDA graphs but against my expectations I am seeing virtually no performance difference compared to before:

| Model                      | GPU         | Static lookup cache | Slots | Seed | Previous runs | Iterations master | Iterations PR --draft 5 | Speedup vs. master |
|----------------------------|-------------|---------------------|-------|------|---------------|-------------------|-------------------------|--------------------|
| LLaMA 3 Instruct 8b q4_K_M | 1x RTX 4090 | None                |     1 |   42 |             0 |               167 |                     183 |               1.10 |
| LLaMA 3 Instruct 8b q4_K_M | 1x RTX 4090 | None                |     2 |   42 |             0 |               277 |                     284 |               1.03 |
| LLaMA 3 Instruct 8b q4_K_M | 1x RTX 4090 | None                |     4 |   42 |             0 |               426 |                     363 |               0.85 |
| LLaMA 3 Instruct 8b q4_K_M | 1x RTX 4090 | None                |     8 |   42 |             0 |               540 |                     417 |               0.77 |